backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
spacy[transformers]>=3.6.0
pdfplumber>=0.10.0
python-docx>=0.8.11
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
requests>=2.31.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
